backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
PyPDF2==3.0.1
httpx==0.27.0
reportlab==4.0.7
